fix: 完成工作台搜索部分内容
