+++ src/template/home-portal/oldIndex.scss
@@ -1,0 +1,102 @@
+@import '@hi-ui/core-css/index.scss';
+
+.page--portal {
+  padding: $spacer-9 $spacer-6;
+  min-height: calc(100vh - 64px);
+  box-sizing: border-box;
+  background-color: $gray-lighter;
+}
+
+.portal {
+  &__container {
+    max-width: 1300px;
+    margin: 0 auto;
+
+    .icon-search {
+      color: get-color($palette-hiui, '50');
+      font-size: $font-size-h3;
+    }
+  }
+
+  &__header {
+    display: flex;
+    align-items: center;
+
+    h3 {
+      margin-top: 0;
+      margin-bottom: $spacer-3;
+      font-size: $font-size-large;
+      font-weight: $font-weight-normal;
+    }
+  }
+
+  &__shortcut-list {
+    flex-wrap: wrap;
+    justify-content: space-between;
+
+    .hi-grid__col-8 {
+      &:nth-child(n+4) {
+        margin-top: $spacer-6;
+      }
+    }
+  }
+
+  &__shortcut-item {
+    .hi-card {
+      height: 120px;
+      box-sizing: border-box;
+      background-color: $white;
+      color: $gray-darker;
+
+      h4 {
+        margin-top: 0;
+        margin-bottom: $spacer-3;
+        font-size: $font-size-large;
+        font-weight: $font-weight-normal;
+        color: $black;
+      }
+
+      p {
+        margin: 0;
+        font-size: $font-size-normal;
+      }
+    }
+  }
+
+  &__todo-item {
+    margin-bottom: $spacer-3;
+
+    & + & {
+      margin-top: 0;
+    }
+  }
+}
+
+.portal-todo-item {
+  justify-content: space-between;
+  font-size: $font-size-normal;
+  line-height: $line-height-size-normal;
+
+  &__title {
+    margin: 0;
+    font-size: $font-size-normal;
+    font-weight: $font-weight-normal;
+    line-height: $line-height-size-normal;
+
+    &,
+    a {
+      color: $black;
+    }
+  }
+
+  &__desc,
+  &__date {
+    font-size: $font-size-small;
+    line-height: $line-height-size-small;
+    color: $gray-dark;
+  }
+
+  &__action {
+    cursor: pointer;
+  }
+}
